disp: msm: shd: move SHD data structures to state
Previously, SHD stores the HW resource allocated to the
shared display into the SHD object. The HW resources will
be reserved in the atomic check stage, and used by the
atomic commit stage to program the hardware.

But if there is mode change request, either because the
base display or shared display mode is changed or a NULL
commit is triggered for feature like Always-Connected-DP,
the atomic check thread might touch the HW resource that is
still using by the atomic commit thread, result inconsistent
configuration of the HW blocks, especially the CTRL_FLUSH
bitmask for the SSPP associated with the SHD, might end up
with SMMU fault.

Move SHD data structure storing the HW resource into the DRM
private state, so that the atomic check and atomic commit
stage would work on different state object and don't
interfere each other. The state will be swapped after atomic
check and before submitting to commit thread.

Change-Id: Ica16aa308ff1949139b4c1e7db2f7686172b2246
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>