msm-hyp : fix for kernel panic issue in case of batch commit
Issue : When batch packet size exceeds the currently allocated size,
memory is extended to current size + 4K, but the condition to
judge whether the packet size exceeds the currently allocated
is wrong, resulting in right red zone rewritten,
the last execution of the wire_port_send_recv judged the need
to expand the memory, call krealloc panic.

Fix : Corrected the condition to judge whether addition of new
packet to the buffer exceeds the currently allocated size or not.

Change-Id: I970987b4e46d5c0513d706b6514693f73016ad43
Signed-off-by: Harish R <quic_harisr@quicinc.com>