disp: msm: dsi: load DSI prop only if selected
DSI prop is always attempted to be loaded when no panel
is selected. This results in a wait for firmware to be loaded, which
can be avoided when DSI prop isn't being used.

This change attempts to load firmware only if UEFI explicitly tells
driver to use DSI prop.

Change-Id: I6696e3607f1f06ee0421d0fa09a6b49495c996f6
Signed-off-by: Kirill Shpin <quic_kshpin@quicinc.com>
(cherry picked from commit 67760b1238e3587c2ad942147af2f482823eb911)