disp: msm: sde: decrease pending kickoff count when skip wait vblank
In the SHD use case, shared display and base display use the same
physics display, if the base display disabled, it would stop the
irqs and clks, shared display wait for vblank will timeout, for
this use case, shared display skip wait for vblank. But the pending
kickoff count only increases and not decrease, that causes the
pending kickoff count always increases, if the count more than
three and the base display re-enabled, the next commit need to wait
for vblank and will timeout 84ms.

Change-Id: I1c59b1e45b0cda5ab5e8ba09038c63833f54f5fc
Signed-off-by: huiyang <quic_huiyang@quicinc.com>