disp: msm: add mutex lock to protect the drm client init
If drm release is called during the drm client init, at
that time the crtc has not attached to the drm client,
then if use this drm client a null pointer issue will
occur. Add mutex lock to protect the drm client init,
it can avoid this kind of issue.

Change-Id: I39f874df8bf87d25d9cc796bd2bd8ee14939e593
Signed-off-by: huiyang <quic_huiyang@quicinc.com>