msm: disp: dp: Avoid HW programming if the DP state is not enabled
In certain cases, the DPCD register may be accessed during DP
unprepare, even when the DP state is not enabled. This could lead
to an SError occurrence.

To prevent this, add a condition to check the DP state if it is
not enabled, avoid any hardware programming.

Change-Id: Ic1d8c26f2ce9c353010e3277b4234172dc66d1ff
Signed-off-by: huiyang <quic_huiyang@quicinc.com>
(cherry picked from commit 7d1ff4b985f3b79e590193f2064617f226054645)