disp: msm: sde: add mutex lock to protect the state cache
When destroy msm prop, no protection for free the state cache.
If another thread accesses the state cache, it can't get the
correct value, add mutex lock to protect the state cache.

Add some conditional checks to increase the robustness of the
code. Set it to a null pointer after freeing the state cache.

Change-Id: Ic8a41d883454f2f1eb58ccbf1f00fc9da3632352
Signed-off-by: huiyang <quic_huiyang@quicinc.com>