disp: msm: sde: if EDEADLK occurred retry commit from the beginning
In the multithread use case, while two thread commit to the same crtc,
if EDEADLK occurred on one thread, it will backoff and drop all the
mode set locks, another thread can get the mode set lock, it will
change the crtc state and plane state. when another thread unlock
the mode set locks, this thread can get the mode set lock and continue
to execute, but the crtc state and plane state are changed by another
thread. Then the crtc state and plane state are incorrect, it also
can cause framebuffer ref count decrease to zero. Then the framebuffer
will be freed and commit failure.

Modify the code logic, if EDEADLK occurred return error and retry
commit from the beginning, it will clear the old state and duplicate a
new state and commit again. That can avoid the commit failure issue.

If a deadlock occurs when sde_rm tries to acquire atomic_state, and it
happens during the atomic check phase, follow the atomic commit flow:
simply return -EDEADLOCK and allow the DRM framework to retry the
operation from the beginning.

Change-Id: I618eb355a3460e0ccebc72458c8ff225c48d0256
Signed-off-by: huiyang <quic_huiyang@quicinc.com>